Created a web page with a 16*16 grid square div
By created a page where 16 box in a row.And each single row has 16 columns.

This a sketch pad so for the pad screen we have selected 16 * 16 screen.So
anyone will be draw over here.

For creating this screen by defining width 800px and height 800px of div.
We have added 16 box in the row,each box with width 25px.In css by add
flexbox to the parent of the box.16*16 div box created with js nested
for loop and add a classlist .box with both 25px width and height.