Divide the sketch pad screen into custom size
By input a number in the field user can create that number of little
screen for the sketch pad.

Rather than fixed size screen by adjusting custom screen user has the
ability to draw large or small object.

When user fill the box with number then click genarate this button
will create the number of box in the both row and column.As our
total screen width and height are fixed we have to change the little
box so they can fit in the large screen.