Change the color of sketch pad as the color picked
When a user pick a color from color picker.The picked color will
apply to the sketchpad box.

Rather than random color and fixed color the user will be able to
picked a specific color.

When user choose a color by select a color from color picker.We will
add a class to the color picker and if the class is present the color
will be changed as the user specified.