Reset button to reset the sketch pad and Black button color the pad black
Add a reset button so when the user click on it,it will clear the
workspace.The black color button will set the color of sketchpad
black.

When the user feel that they want to clear the sketch pad this
button will came handy.The the user set the color as random if
they want to color again black help them with that.

Added class for both reset and black button.When user click on them
reset button remove all the box.Black button will color the box
black.